APIS/backend/requirements.txt
asgiref==3.5.2
Django
django-extensions==3.2.1
django-filter==22.1
djangorestframework
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
pinecone-client
OpenAI
langchain
llama_index==0.6.23
transformers
django-cors-headers
